requirements.txt
feedparser==6.0.12
growwapi==1.3.0
openai==2.8.1
pandas==2.3.3
pymongo==4.15.5
pyotp==2.9.0
python-dotenv==1.2.1
